phase 6: ApplicationRepository relocation + SDK execution-path migration
Repository relocation:
- Moved ApplicationRepository from routers/applications.py (518 lines)
  to repositories/applications.py. Per repositories/README.md the
  repository pattern lives in repositories/, not in router files.
  Router now imports from the new location.

SDK execution-path migration:
- cli_list_tables: replaced inline cascade with TableRepository.list().
  The base class handles cascade (org + global) via OrgScopedRepository
  semantics. This is the SDK execution path — same pattern as Phase 4
  did for OAuth.
- cli_create_table: documented why the exact-scope uniqueness check is
  NOT a cascade. Allow-list entry updated to call this out.

Scope decision (deviation from original plan):
The plan called for migrating every inline cascade across all routers
(~47 callsites). Phase 6 focused on the SDK execution path (cli.py)
and the documented-pattern-compliance work (ApplicationRepository
relocation). UI-facing router migrations are deferred to Phase 8
follow-up issues — they're tracked in the lint allow-list as the
permanent record. The plan section is updated to reflect.

Allow-list shrinkage (the visible signal of migration progress):
- ApplicationRepository cascade-override entries: 2 removed
- cli_list_tables inline cascade entries: 2 removed
- cli_create_table entry: justification updated (NOT cascade)

Tests: 3974 pass. The lint test caught the now-stale allow-list entries
automatically — Phase 3's mechanical enforcement is doing its job.

Pyright + ruff clean on touched files.
